+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 **unreleased**
+
+**v0.0.61**
 - [Feature] Move files in `GcsHook` using batch operations to copy and delete files in order to improve performance
 - [Feature] Add random hash to filename to increase upload velocity
 - [Feature] Add logic to control which file timestamps are being processed